RDKB-34882 : Debugging "no such file or directory"- eblist
Reason for change: To eliminate "No such file or directory-rm /tmp/eblist" error
Test Procedure: N/A
Risks: None

(cherry picked from commit 5d646d3ea26981bcd473068dddb7599ed2ee0625)

+++ scripts/rxtx_res.sh
@@ -20,5 +20,9 @@
 #reset and stop the count 
 #ebtables -F;
 traffic_count -F;
-rm /tmp/eblist;
+if [ -f /tmp/eblist ];then
+	rm /tmp/eblist;
+else
+	echo "eblist not found"
+fi
 rm /tmp/rxtx_enabled;
